+++ lib/handlers/check-order-status/util.ts
@@ -175,3 +175,16 @@ export function getValidator(provider: ethers.providers.StaticJsonRpcProvider, c
   }
   return validatorMap.get(chainId) as OrderValidator
 }
+
+/*
+ * In the first hour of order submission, we check the order status roughly every block.
+ * We then do exponential backoff on the wait time until the interval reaches roughly 6 hours.
+ * All subsequent retries are at 6 hour intervals.
+ */
+export function calculateDutchRetryWaitSeconds(chainId: ChainId, retryCount: number): number {
+  return retryCount <= 300
+    ? AVERAGE_BLOCK_TIME(chainId)
+    : retryCount <= 450
+    ? Math.ceil(AVERAGE_BLOCK_TIME(chainId) * Math.pow(1.05, retryCount - 300))
+    : 18000
+}
