+++ test/integ/handlers/check-order-status-service.test.ts
@@ -69,9 +69,9 @@ describe('checkOrderStatusService', () => {
       checkOrderStatusService = new CheckOrderStatusService(
         ordersRepositoryMock,
         OrderType.Dutch,
+        analyticsMock,
         FILL_EVENT_LOOKBACK_BLOCKS_ON,
-        calculateDutchRetryWaitSeconds,
-        analyticsMock
+        calculateDutchRetryWaitSeconds
       )
 
       watcherMock = {
@@ -379,33 +379,6 @@ describe('checkOrderStatusService', () => {
           })
         )
       })
-    })
-  })
-
-  describe('CheckOrderStatusService.calculateRetryWaitSeconds', () => {
-    let ordersRepositoryMock: any, checkOrderStatusService: CheckOrderStatusService
-
-    beforeEach(() => {
-      ordersRepositoryMock = {
-        updateOrderStatus: jest.fn(),
-        getByHash: jest.fn(),
-      } as any
-      checkOrderStatusService = new CheckOrderStatusService(ordersRepositoryMock, OrderType.Dutch)
-    })
-
-    it('should do exponential backoff when retry count > 300', async () => {
-      const response = calculateDutchRetryWaitSeconds(1, 301)
-      expect(response).toEqual(13)
-    })
-
-    it('should do exponential backoff when retry count > 300', async () => {
-      const response = calculateDutchRetryWaitSeconds(1, 350)
-      expect(response).toEqual(138)
-    })
-
-    it('should cap exponential backoff when wait interval reaches 18000 seconds', async () => {
-      const response = calculateDutchRetryWaitSeconds(1, 501)
-      expect(response).toEqual(18000)
     })
   })
 })
